+++ tests/property/krylov/mod.rs
@@ -3,4 +3,5 @@
 #[allow(dead_code)]
 mod common;
 mod inverse_power_iteration;
+mod lanczos;
 mod power_iteration;